addressing fixes and adding a few things

+++ taskmaster/views.py
@@ -54,10 +54,8 @@ def show_queue_form():
     return render_template('create_queue.html')
 
 
-@app.route('/queue', methods=['POST'])
-def create_queue():
-    if request.method == 'POST':
-        name  = request.form['queue-name']
+@app.route('/queue/<name>', methods=['POST'])
+def create_queue(name):
     db.create_queue(name, org)
 
     return render_template('create_queue.html')
@@ -68,14 +66,15 @@ def task_tags(task_id):
 
     return Response(status=200)
 
-@app.route('/queue/delete/<queue_name>', methods=['POST'])
-def delete_queue(queue_name):
-    db.delete_queue(queue_name, org)
+@app.route('/queue/<queue_name>', methods=['DELETE'])
+def update_queue(queue_name):
+    if request.method == 'DELETE':
+        db.delete_queue(queue_name, org)
 
-    return Response(status=200)
+        return Response(status=200)
 
-@app.route('/task/<task_id>/update_status/<status>', methods=['POST'])
-def update_task_status(task_id, status):
-    db.update_task_status(task_id, status)
+@app.route('/task/<task_id>/update/<update_field>/<update_value>', methods=['POST'])
+def update_task(task_id, update_field, update_value):
+    db.update_task(task_id, update_field, update_value)
 
     return Response(status=200)
